Abstracts out Message to its own module
